Implement WASI's fd_datasync
